Disable delete key back-navigation in Cocoa
Prior to this commit, pressing the Delete key on the Mac keyboard caused
the WebView to navigate to the previous HTML page, which is undesirable
in a desktop application.

This fix makes use of the WebView's policyDelegate to refuse a back-
navigation actions if it was triggered by pressing the Delete key.
This is most convinient for two reasons:

    1. We need not worry about Delete key press events that may occur
at other places (say inside text-fields).
    2. Back navigation triggered explicitly via other means (like the
JavaScript history.back() method) can be allowed gracefully.

Note: attempts to capture the keypress via the WebView's keyDown method
fails because the parent NSWindow automatically assigns a subview called
WebHTMLView as the firstResponder to keyDown events, which handles
Delete key press events itself.