[GTK] Return result of evaluated JS code
In WebkitGTK+, the evaluate_script() function is defined to return void.
So inorder to obtain the result of the executed script, we need to take
the indirect approach of evaluating the script using eval(), storing the
result in the document title, and retrieving it later by listening to
the notify::title signal.