[Cocoa] Add a unique instance id to each window
To distinguish multiple instances, all new instances are now assigned a
random UUID, which is returned by create_window to the caller.

Since the very first call to create_window blocks program execution and
cannot return, the initial window is assigned the special UUID 'master'.

* Add parameter uid to API functions

Default value of the parameter is 'master'. So single window apps can
omit this, and we remain backwards-compatible.

The added get_instance method is able to handle any attribute so that it
can be used elsewhere. e.g. this avoids repitition of code that loops
through the instances list.