[GTK] Implement basic JS-Python bridge support
PyGTK does not support exposing objects to JavaScript. So it's not
possible to hook the `call` methods on either side of the bridge.

The current approach is to write the function name and parameter strings
to the `status` attribute of the JS window; this triggers a signal which
is handled to retrieve the strings and invoke `call` manually. A unique
id assigned to the JSBridge object is used to distinguish status updates
triggered in this way.