[Qt 5.5+] Implement basic web inspector support
Starting with 5.5, Qt relies on Chrome Remote debugging for inspecting
webviews. This commit utilises the multi-window feature to open a new
window, using it to host the debugger instead of an external Chromium
based browser.

The info about the debug server is automatically logged by Qt into the
console, so users who prefer to debug in an external browser can do so.

Note: this has not passed the experimental stage; it has still got a few
bugs to take care of.