Individual windows will now start a server only if their URL is a web app. All other traffic goes to the global server.
Modified multiple_servers example to call start with http_server=True since
the webapps will no longer cause the global server to be started. Also added
some global server detection to the example so if http_server is set False the
third window will correctly display that the global server is not started, instead
of crashing.