Fix ever-growing main menu bug
Adding the default menu items on every `first_show()` results in an ever-growing main menu, very
visible during unit test runs.

Instead of blindly adding the default items every time, make sure to reset the main menu to an
empty state, as it would be on the first call to `first_show()`.